+++ RELEASE_NOTES.md
@@ -2,13 +2,15 @@
 
 ## v2.0.0 (TBD)
 
+* Added `#getRefDetails` to centralize the code used to generate reference metadata *(Also allows you to see why an
+object you expect to be returned by `#findRefs` is not returned.)*
 * Renamed `#isJsonPointer` to `#isJsonPtr`
 * Renamed `#isJsonReference` to `#isJsonRef`
 * Renamed `#pathFromPointer` to `#pathFromPtr`
 * Renamed `#pathToPointer` to `#pathToPtr`
 * Updated `#findRefs` to no longer use [js-traverse](https://github.com/substack/js-traverse), to no longer process
 child properties of JSON Reference objects and to use an *options* object
-* Updated `#isJsonPtr` to valdiate the `$ref` value is a URI instead of treating all string values as valid
+* Updated `#isJsonPtr` to validate the `$ref` value is a URI instead of treating all string values as valid
 * Updated `#pathToPtr` to take an optional second argment to allow for returning both hash-based *(default)* and
 slash-based JSON Pointers